+++ CHANGELOG.md
@@ -7,10 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [Unreleased]
+## [1.1.6]
 
 ### Changed
-- Allow puma 8
+- Allow puma 8 (#42)
 
 ## [1.1.5]